pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "codecoverage-cli"
version = "0.1.0"
description = "AI-powered test generation and documentation for Python codebases"
readme = "README.md"
requires-python = ">=3.10"

authors = [
    {name = "Abhigya Shridhar", email = "abhigyashridhar101@gmail.com"},
]
license = {text = "MIT"}

keywords = [
    "testing",
    "ai",
    "code-generation",
    "llm",
    "langgraph",
    "rag",
]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Code Generators",
]

# ============================================================================
# DEPENDENCIES
# ============================================================================

dependencies = [
    "click>=8.1.0",           # CLI framework
    "rich>=13.0.0",           # Beautiful terminal output

    "pydantic>=2.0.0",        # Data validation using Python type hints
    "pydantic-settings>=2.0.0", # Environment variable loading
    "pyyaml>=6.0",            # YAML config file parsing

    "astroid>=3.0.0",         # Advanced AST analysis
    "radon>=6.0.0",           # Code complexity metrics
    "gitignore-parser>=0.1.0", # Parse .gitignore files

    "anthropic>=0.18.0",      # Claude API client (official)
    "langgraph>=0.0.40",      # Agent workflow orchestration
    "langchain-core>=0.1.0",  # Core LangChain abstractions
    "langchain-anthropic>=0.1.0", # Anthropic integration for LangChain

    "gitpython>=3.1.0",       # Git operations from Python

    "tenacity>=8.2.0",        # Retry logic with exponential backoff
    "jinja2>=3.1.0",          # Template engine for prompts

    "tomli>=2.0.0; python_version < '3.11'",  # TOML parsing for older Python
]

[project.optional-dependencies]

dev = [
    "pytest>=7.4.0",          # Testing framework
    "pytest-cov>=4.1.0",      # Coverage reporting
    "pytest-asyncio>=0.21.0", # Async test support (for LangGraph)
    "black>=23.0.0",          # Code formatter (opinionated)
    "ruff>=0.1.0",            # Fast linter (replaces flake8, isort)
    "mypy>=1.7.0",            # Static type checker
    "ipython>=8.0.0",         # Better Python REPL
    "ipdb>=0.13.0",           # IPython debugger
]

openai = [
    "langchain-openai>=0.1.0",  # OpenAI provider support
]

all = [
    "codecoverage[dev,openai]",
]

[project.scripts]
codecoverage = "codecoverage.cli.main:cli"

[project.urls]
Homepage = "https://github.com/AbhigyaShridhar/codecoverage-cli/"
Documentation = "https://github.com/AbhigyaShridhar/codecoverage-cli#readme"
Repository = "https://github.com/AbhigyaShridhar/codecoverage-cli/"
Issues = "https://github.com/AbhigyaShridhar/codecoverage-cli/issues"
Changelog = "https://github.com/AbhigyaShridhar/codecoverage-cli/releases"

[tool.hatch.build.targets.wheel]
packages = ["src/codecoverage"]

[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py310"

select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "PT",  # flake8-pytest-style
]

ignore = [
    "E501",  # Line too long (black handles this)
]

# Enforce specific exception handling
[tool.ruff.per-file-ignores]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true

[[tool.mypy.overrides]]
module = [
    "rank_bm25.*",
    "radon.*",
    "gitignore_parser.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
]
# Run with coverage:  pytest --cov=codecoverage --cov-report=term-missing

[tool.coverage.run]
branch = true
source = ["codecoverage"]
omit = [
    "*/tests/*",
    "*/__main__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]